Default Commit Message

+++ src/gyro.rs
@@ -16,7 +16,7 @@ pub fn init_gyro(channel: Sender<serde_json::Value>, timer: Arc<howlong::HighRes
             let gyro_readings = gyro.read();
             channel.send(serde_json::json!({
                 "gyroscope": gyro_readings,
-                "time": timer.elapsed().as_nanos()
+                "time": timer.elapsed().as_nanos() as u64
             })).unwrap();
         }
     });
